+++ engine/src/check.hpp
@@ -14,7 +14,7 @@ namespace wisdom
     // check if the the king indicated by the WHO argument is in trouble
     // in this position
     bool is_king_threatened (const Board& board, Color who, int8_t row, int8_t col);
-    bool is_king_threatened_inline (const Board& board, Color who, int row, int col);
+    bool is_king_threatened_inline (const Board& board, Color who, int8_t row, int8_t col);
 
     bool is_king_threatened_row (const Board& board, Color who, int8_t row, int8_t col);
     bool is_king_threatened_column (const Board& board, Color who, int8_t row, int8_t col);
